backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.10.4
    # via flask-migrate
aniso8601==10.0.0
    # via flask-restx
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.2.1
    # via flask-bcrypt
blinker==1.6.2
    # via flask
click==8.1.3
    # via flask
flasgger==0.9.7.1
    # via -r requirements.in
flask==2.3.2
    # via
    #   -r requirements.in
    #   flasgger
    #   flask-bcrypt
    #   flask-cors
    #   flask-login
    #   flask-migrate
    #   flask-restx
    #   flask-sqlalchemy
    #   flask-wtf
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-cors==4.0.0
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-migrate==4.0.4
    # via -r requirements.in
flask-restx==1.3.0
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via flask-migrate
flask-wtf==1.2.2
    # via -r requirements.in
gevent==24.11.1
    # via -r requirements.in
greenlet==3.1.1
    # via
    #   gevent
    #   sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
importlib-resources==6.5.2
    # via flask-restx
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
jsonschema==4.23.0
    # via
    #   flasgger
    #   flask-restx
jsonschema-specifications==2024.10.1
    # via jsonschema
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mistune==3.1.0
    # via flasgger
packaging==24.2
    # via
    #   flasgger
    #   gunicorn
psycopg2==2.9.10
    # via -r requirements.in
pytz==2024.2
    # via flask-restx
pyjwt==2.10.1
    # via -r requirements.in
pytest==8.3.4
    # via -r requirements.in
pyyaml==6.0.2
    # via flasgger
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via flasgger
sqlalchemy==2.0.12
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   alembic
    #   mistune
    #   sqlalchemy
werkzeug==2.3.3
    # via
    #   flask
    #   flask-login
    #   flask-restx
wtforms==3.2.1
    # via flask-wtf
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
